+++ src/main/java/frc/robot/RobotContainer.java
@@ -13,7 +13,7 @@ import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 
 import frc.robot.shuffleboard.GRTShuffleboardTab;
-import frc.robot.subsystems.tank.Tank;
+import frc.robot.subsystems.tank.TankSubsystem;
 import java.lang.Math;
 
 /**
@@ -26,7 +26,7 @@ import java.lang.Math;
 public class RobotContainer {
     // Subsystems
     //private final SwerveSubsystem swerveSubsystem;
-    final Tank tank = new Tank();
+    final TankSubsystem tank = new TankSubsystem();
     private final XboxController controller = new XboxController(0);
 
     // Controllers and buttons
